requirements.txt
# Gemini API
google-genai>=0.4.0

# Audio handling
pyaudio
pyalsaaudio
numpy

# Environment variables
python-dotenv

# Gmail API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# GPIO (Raspberry Pi)
gpiozero
# lgpio - システムのpython3-lgpioを使用

# HTTP requests (Firebase)
requests
